update import script for new schema
