+++ import.py
@@ -5,6 +5,8 @@ import logging
 import os
 import sqlite3
 
+from utils import db_connect, migrate_db
+
 
 parser = argparse.ArgumentParser()
 parser.add_argument('directory', help=(
@@ -20,19 +22,16 @@ assert log_level in ['CRITICAL', 'ERROR', 'WARNING', 'INFO', 'DEBUG']
 logging.basicConfig(level=getattr(logging, log_level))
 logger = logging.getLogger(__name__)
 
-conn = sqlite3.connect(args.database_path)
-cursor = conn.cursor()
-cursor.execute('create table if not exists messages (message text, user text, channel text, timestamp text, UNIQUE(channel, timestamp) ON CONFLICT REPLACE)')
-cursor.execute('create table if not exists users (name text, id text, avatar text, UNIQUE(id) ON CONFLICT REPLACE)')
-cursor.execute('create table if not exists channels (name text, id text, UNIQUE(id) ON CONFLICT REPLACE)')
+conn, cursor = db_connect(args.database_path)
+migrate_db(conn, cursor)
 
 directory = args.directory
 
 logger.info("Importing channels..")
 with open(os.path.join(directory, 'channels.json')) as f:
     channels = json.load(f)
-args = [(c['name'], c['id']) for c in channels]
-cursor.executemany('INSERT INTO channels VALUES(?,?)', (args))
+args = [(c['name'], c['id'], 1) for c in channels]
+cursor.executemany('INSERT INTO channels VALUES(?,?,?)', (args))
 logger.info("- Channels imported")
 
 logger.info("Importing users..")
